fix(android): Online-button crash (curl int/long mismatch) + launcher login URL
Two independent bugs found from real device testing:

1. Native crash on pressing "Online" in-game. Static-analysis investigation
   (via a dedicated research agent, no device log was available yet) found
   the root cause: CURLINFO_RESPONSE_CODE requires a `long*` per libcurl's
   documented API, but this module declared `int httpResponseCode` in three
   places (OnlineServices_RoomsInterface.cpp's WebSocket::Connect() and
   ::Tick(), HTTPRequest.cpp's m_responseCode member) and passed its address
   directly. On LP64 platforms (Android arm64) `long` is 8 bytes vs `int`'s
   4 -- curl writes a full 8-byte value through a 4-byte pointer, corrupting
   adjacent stack/member memory. This is the very first curl call made in
   the whole GeneralsOnline connect path, executed unconditionally and
   synchronously on every "Online" press, and explains why it's a new bug:
   on Windows (LLP64) long==int==4 bytes, so the identical code never
   overflowed there. Fixed all three to `long`, updated the one printf
   format specifier that referenced m_responseCode (%d -> %ld), and fixed
   the same int/long mismatch on four curl_easy_setopt calls that take
   `long`-typed options (CURLOPT_HTTP_VERSION x2, CURLOPT_CONNECTTIMEOUT_MS,
   CURLOPT_TIMEOUT_MS, CURLOPT_PROXYPORT) but were passed plain `int`/
   `uint16_t` expressions -- same class of variadic-call UB, lower
   confidence these specifically crash but they're the identical mistake.

2. GeneralsOnlineActivity.java's login URL was missing the reference
   client's &client=<id> query param (OnlineServices_Auth.cpp's BeginLogin
   always includes it) -- without it, the site apparently doesn't reliably
   associate the generated code with the pending login: the browser says
   "return to the game" but CheckLogin polling never resolves, leaving the
   launcher stuck on "Not signed in" with a network-error toast until
   POLL_MAX_ATTEMPTS gives up.

+++ GeneralsMD/Code/GameEngine/Include/GameNetwork/GeneralsOnline/HTTP/HTTPRequest.h
@@ -65,7 +65,10 @@ private:
 private:
 	CURL* m_pCURL = nullptr;
 
-	int m_responseCode = -1;
+	// GeneralsX @bugfix Android port 10/07/2026 CURLINFO_RESPONSE_CODE needs
+	// a long*, not int* -- see OnlineServices_RoomsInterface.cpp for the
+	// same fix and the full explanation (LP64 stack/member corruption).
+	long m_responseCode = -1;
 
 	EHTTPVerb m_httpVerb;
